Corrected x64 version

+++ casterdlg.cpp
@@ -287,7 +287,7 @@ void astRender(CSkPainter *p, mapView_t *view, float maxMag)
           setSetFont(FONT_ASTER, p);
           p->drawTextLR(pt.sx + size + 1, pt.sy + size + 1, a->name);
         }
-        addMapObj(pt.sx, pt.sy, MO_ASTER, MO_CIRCLE, size + 2, i, (int)a, a->orbit.mag);
+        addMapObj(pt.sx, pt.sy, MO_ASTER, MO_CIRCLE, size + 2, i, (qint64)a, a->orbit.mag);
       }
     }
   }
